Fix + rename CancelReservationController > ReservationCancellationController
