+++ HetDepot/ExampleFile/ExampleSettings.json
@@ -56,13 +56,17 @@
     "tourOneReservation": "{count} reservering",
     "tourReservations": "{count} reserveringen",
     "visitorAlreadyHasTourAdmission": "U heeft al deelgenomen aan een rondleiding vandaag. Morgen weer een kans.",
+    "visitorEnterCode": "Vul uw code in:",
     "visitorLogonCodeInvalid": "De code is niet geldig. Controleer uw code en probeer het nog eens.",
     "visitorRequestCodeSelectedTour": "U wilt zich aanmelden voor de rondleiding van {time}",
     "visitorRequestCodeSelectedTourForGroup": "Persoon in gezelschap aanmelden voor de rondleiding van {time}",
     "visitorReservationAlreadyHavingOne": "U heeft al een reservering voor dit tijdstip. Wilt u deze annuleren?.",
     "visitorReservationAlreadyHavingOneForGroup": "De persoon die u aan wilt melden heeft al een reservering voor dit tijdstip.",
+    "visitorReservationCancellationAlreadyUsed": "Deze code is al gebruikt voor een rondleiding",
     "visitorReservationCancellationConfirmation": "Reservering geannuleerd",
+    "visitorReservationCancellationNoReservation": "Er is geen reservering voor deze code",
     "visitorReservationCancellationRequestionConfirmation": "Wilt u deze reservering annuleren?",
+    "visitorReservationCancellationTitle": "Reservering annuleren",
     "visitorReservationChangeNotChanges": "De reservering voor {time} blijft behouden",
     "visitorReservationChangeTourConfirmation": "U bent uitgeschreven voor {currentTime}. Uw nieuwe rondleiding start om {newTime}",
     "visitorReservationChangeTourConfirmationForGroup": "Persoon uitgeschreven voor {currentTime}. De nieuwe rondleiding start om {newTime}",
